mv tests/conftest.py crystal_toolkit/apps/examples/tests/typing.py
